This commit was generated by cvs2svn to compensate for changes in r12043, which included commits to RCS files with non-trunk default branches.
git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@12044 379a1393-f5fb-40a0-bcee-ef074d9b53f7

+++ vice/src/sounddrv/soundwav.c
@@ -64,10 +64,10 @@ static int wav_init(const char *param, int *speed,
     samples = 0;
 
     /* Initialize header. */
-    le_store(header + 22, *channels, 2);
+    le_store(header + 22, (DWORD)*channels, 2);
     le_store(header + 24, sample_rate, 4);
     le_store(header + 28, bytes_per_sec, 4);
-    le_store(header + 32, *channels*2, 2);
+    le_store(header + 32, (DWORD)*channels*2, 2);
 
     return (fwrite(header, 1, 44, wav_fd) != 44);
 }
